refactor the code by creating photosTemplate to reduce duplicity
